+++ packages/col/index.ts
@@ -15,22 +15,12 @@ VantComponent({
     style: ''
   },
 
-  computed: {
-    classes(): string {
-      const { span, offset } = this.data;
-      return this.classNames('van-col', {
-        [`van-col--${span}`]: span,
-        [`van-col--offset-${offset}`]: offset
-      });
-    }
-  },
-
   methods: {
     setGutter(gutter: number) {
       const padding = `${gutter / 2}px`;
       const style = gutter ? `padding-left: ${padding}; padding-right: ${padding};` : '';
       if (style !== this.data.style) {
-        this.setData({ style });
+        this.set({ style });
       }
     }
   }