+++ packages/dialog/dialog.ts
@@ -9,7 +9,9 @@ type DialogOptions = {
   message?: string;
   overlay?: boolean;
   selector?: string;
+  transition?: string;
   asyncClose?: boolean;
+  messageAlign?: string;
   confirmButtonText?: string;
   cancelButtonText?: string;
   showConfirmButton?: boolean;
@@ -37,13 +39,18 @@ function getContext() {
 }
 
 const Dialog: Dialog = options => {
+  options = {
+    ...Dialog.currentOptions,
+    ...options
+  };
+
   return new Promise((resolve, reject) => {
     const context = options.context || getContext();
     const dialog = context.selectComponent(options.selector);
     delete options.selector;
 
     if (dialog) {
-      dialog.setData({
+      dialog.set({
         onCancel: reject,
         onConfirm: resolve,
         ...options
@@ -62,6 +69,8 @@ Dialog.defaultOptions = {
   zIndex: 100,
   overlay: true,
   asyncClose: false,
+  messageAlign: '',
+  transition: 'scale',
   selector: '#van-dialog',
   confirmButtonText: '确认',
   cancelButtonText: '取消',
@@ -71,15 +80,10 @@ Dialog.defaultOptions = {
   confirmButtonOpenType: ''
 };
 
-Dialog.alert = options =>
-  Dialog({
-    ...Dialog.currentOptions,
-    ...options
-  });
+Dialog.alert = Dialog;
 
 Dialog.confirm = options =>
   Dialog({
-    ...Dialog.currentOptions,
     showCancelButton: true,
     ...options
   });