+++ packages/dialog/index.ts
@@ -10,6 +10,7 @@ VantComponent({
     message: String,
     useSlot: Boolean,
     asyncClose: Boolean,
+    messageAlign: String,
     showCancelButton: Boolean,
     closeOnClickOverlay: Boolean,
     confirmButtonOpenType: String,
@@ -32,6 +33,10 @@ VantComponent({
     overlay: {
       type: Boolean,
       value: true
+    },
+    transition: {
+      type: String,
+      value: 'scale'
     }
   },
 
@@ -63,7 +68,7 @@ VantComponent({
 
     handleAction(action) {
       if (this.data.asyncClose) {
-        this.setData({
+        this.set({
           [`loading.${action}`]: true
         });
       }
@@ -72,13 +77,13 @@ VantComponent({
     },
 
     close() {
-      this.setData({
+      this.set({
         show: false
       });
     },
 
     stopLoading() {
-      this.setData({
+      this.set({
         loading: {
           confirm: false,
           cancel: false
@@ -91,7 +96,9 @@ VantComponent({
         this.close();
       }
       this.$emit('close', action);
-      this.$emit(action);
+      
+      //把 dialog 实例传递出去，可以通过 stopLoading() 在外部关闭按钮的 loading
+      this.$emit(action, { dialog: this });
 
       const callback = this.data[action === 'confirm' ? 'onConfirm' : 'onCancel'];
       if (callback) {