+++ packages/mixins/transition.ts
@@ -16,28 +16,43 @@ export const transition = function(showDefaultValue) {
     data: {
       type: '',
       inited: false,
-      display: false
+      display: false,
+      supportAnimation: true
     },
 
     attached() {
       if (this.data.show) {
         this.show();
       }
+
+      this.detectSupport();
     },
 
     methods: {
+      detectSupport() {
+        wx.getSystemInfo({
+          success: info => {
+            if (info && info.system && info.system.indexOf('iOS 8') === 0) {
+              this.set({ supportAnimation: false });
+            }
+          }
+        });
+      },
+
       observeShow(value) {
         if (value) {
           this.show();
         } else {
-          this.setData({
-            type: 'leave'
-          });
+          if (this.data.supportAnimation) {
+            this.set({ type: 'leave' });
+          } else {
+            this.set({ display: false });
+          }
         }
       },
 
       show() {
-        this.setData({
+        this.set({
           inited: true,
           display: true,
           type: 'enter'
@@ -46,7 +61,7 @@ export const transition = function(showDefaultValue) {
 
       onAnimationEnd() {
         if (!this.data.show) {
-          this.setData({
+          this.set({
             display: false
           });
         }