+++ packages/slider/index.ts
@@ -6,6 +6,9 @@ VantComponent({
 
   props: {
     disabled: Boolean,
+    useButtonSlot: Boolean,
+    activeColor: String,
+    inactiveColor: String,
     max: {
       type: Number,
       value: 100
@@ -73,10 +76,12 @@ VantComponent({
     updateValue(value, end) {
       value = this.format(value);
 
-      this.setData({
+      this.set({
         value,
         barStyle: `width: ${value}%; height: ${this.data.barHeight};`
       });
+
+      this.$emit('drag', { value });
 
       if (end) {
         this.$emit('change', value);