+++ packages/submit-bar/index.ts
@@ -1,6 +1,9 @@
 import { VantComponent } from '../common/component';
+import { iphonex } from '../mixins/iphonex';
 
 VantComponent({
+  mixins: [iphonex],
+
   classes: [
     'bar-class',
     'price-class',