+++ packages/tabbar/index.ts
@@ -1,6 +1,9 @@
 import { VantComponent } from '../common/component';
+import { iphonex } from '../mixins/iphonex';
 
 VantComponent({
+  mixins: [iphonex],
+
   relation: {
     name: 'tabbar-item',
     type: 'descendant',
@@ -20,6 +23,7 @@ VantComponent({
 
   props: {
     active: Number,
+    activeColor: String,
     fixed: {
       type: Boolean,
       value: true
@@ -37,19 +41,22 @@ VantComponent({
 
   watch: {
     active(active) {
-      this.setData({ currentActive: active });
+      this.set({ currentActive: active });
       this.setActiveItem();
     }
   },
 
   created() {
-    this.setData({ currentActive: this.data.active });
+    this.set({ currentActive: this.data.active });
   },
 
   methods: {
     setActiveItem() {
       this.data.items.forEach((item, index) => {
-        item.setActive(index === this.data.currentActive);
+        item.setActive({
+          active: index === this.data.currentActive,
+          color: this.data.activeColor
+        });
       });
     },
 
@@ -57,7 +64,7 @@ VantComponent({
       const active = this.data.items.indexOf(child);
       if (active !== this.data.currentActive && active !== -1) {
         this.$emit('change', active);
-        this.setData({ currentActive: active });
+        this.set({ currentActive: active });
         this.setActiveItem();
       }
     }