+++ packages/tree-select/index.ts
@@ -3,6 +3,15 @@ import { VantComponent } from '../common/component';
 const ITEM_HEIGHT = 44;
 
 VantComponent({
+  classes: [
+    'main-item-class',
+    'content-item-class',
+    'main-active-class',
+    'content-active-class',
+    'main-disabled-class',
+    'content-disabled-class'
+  ],
+
   props: {
     items: Array,
     mainActiveIndex: {
@@ -10,8 +19,7 @@ VantComponent({
       value: 0
     },
     activeId: {
-      type: Number,
-      value: 0
+      type: [Number, String]
     },
     maxHeight: {
       type: Number,
@@ -51,14 +59,17 @@ VantComponent({
     // 当一个导航被点击时
     onClickNav(event: Weapp.Event) {
       const { index } = event.currentTarget.dataset;
-      this.$emit('click-nav', { index });
+      const item = this.data.items[index];
+      if (!item.disabled) {
+        this.$emit('click-nav', { index });
+      }
     },
 
     // 更新子项列表
     updateSubItems() {
       const selectedItem = this.data.items[this.data.mainActiveIndex] || {};
 
-      this.setData({ subItems: selectedItem.children || [] });
+      this.set({ subItems: selectedItem.children || [] });
 
       this.updateItemHeight();
     },
@@ -70,12 +81,12 @@ VantComponent({
         this.data.subItems.length * ITEM_HEIGHT
       );
 
-      this.setData({ mainHeight: Math.min(maxHeight, this.data.maxHeight) });
+      this.set({ mainHeight: Math.min(maxHeight, this.data.maxHeight) });
     },
 
     // 更新子项列表高度，根据可展示的最大高度和当前子项列表的高度决定
     updateItemHeight() {
-      this.setData({
+      this.set({
         itemHeight: Math.min(
           this.data.subItems.length * ITEM_HEIGHT,
           this.data.maxHeight